support old functionality when feature flag disabled
